--- .rspec_status
@@ -1,42 +1,0 @@
-example_id                                                       | status | run_time        |
----------------------------------------------------------------- | ------ | --------------- |
-./spec/features/javascript_spec.rb[1:1]                          | failed | 0.01501 seconds |
-./spec/features/javascript_spec.rb[1:2:1]                        | failed | 0.00149 seconds |
-./spec/raygun/breadcrumbs/breadcrumb_spec.rb[1:1:1]              | passed | 0.00149 seconds |
-./spec/raygun/breadcrumbs/breadcrumb_spec.rb[1:1:2]              | passed | 0.00042 seconds |
-./spec/raygun/breadcrumbs/breadcrumb_spec.rb[1:1:3]              | passed | 0.00025 seconds |
-./spec/raygun/breadcrumbs/breadcrumb_spec.rb[1:1:4]              | passed | 0.00028 seconds |
-./spec/raygun/breadcrumbs/breadcrumb_spec.rb[1:1:5]              | passed | 0.00025 seconds |
-./spec/raygun/breadcrumbs/breadcrumb_spec.rb[1:1:6]              | passed | 0.00033 seconds |
-./spec/raygun/breadcrumbs/breadcrumb_spec.rb[1:1:7]              | passed | 0.00024 seconds |
-./spec/raygun/breadcrumbs/breadcrumb_spec.rb[1:1:8]              | passed | 0.00041 seconds |
-./spec/raygun/breadcrumbs/breadcrumb_spec.rb[1:2:1]              | passed | 0.00069 seconds |
-./spec/raygun/breadcrumbs/breadcrumb_spec.rb[1:2:2]              | passed | 0.00038 seconds |
-./spec/raygun/breadcrumbs/breadcrumb_spec.rb[1:2:3]              | passed | 0.00043 seconds |
-./spec/raygun/breadcrumbs/breadcrumb_spec.rb[1:2:4]              | passed | 0.0004 seconds  |
-./spec/raygun/breadcrumbs/breadcrumb_spec.rb[1:2:5]              | passed | 0.00065 seconds |
-./spec/raygun/breadcrumbs/store_spec.rb[1:1:1]                   | passed | 0.00055 seconds |
-./spec/raygun/breadcrumbs/store_spec.rb[1:1:2]                   | passed | 0.00054 seconds |
-./spec/raygun/breadcrumbs/store_spec.rb[1:2:1]                   | passed | 0.00056 seconds |
-./spec/raygun/breadcrumbs/store_spec.rb[1:2:2]                   | passed | 0.00022 seconds |
-./spec/raygun/breadcrumbs/store_spec.rb[1:2:3]                   | passed | 0.00042 seconds |
-./spec/raygun/breadcrumbs/store_spec.rb[1:3:1]                   | passed | 0.00218 seconds |
-./spec/raygun/breadcrumbs/store_spec.rb[1:4:1]                   | passed | 0.00813 seconds |
-./spec/raygun/breadcrumbs/store_spec.rb[1:5:1]                   | passed | 0.00042 seconds |
-./spec/raygun/breadcrumbs/store_spec.rb[1:5:2]                   | passed | 0.00036 seconds |
-./spec/raygun/breadcrumbs/store_spec.rb[1:5:3]                   | passed | 0.0003 seconds  |
-./spec/raygun/breadcrumbs/store_spec.rb[1:5:4]                   | passed | 0.00029 seconds |
-./spec/raygun/breadcrumbs/store_spec.rb[1:5:5]                   | passed | 0.00038 seconds |
-./spec/raygun/breadcrumbs/store_spec.rb[1:5:6]                   | passed | 0.00036 seconds |
-./spec/raygun/breadcrumbs/store_spec.rb[1:5:7]                   | passed | 0.0003 seconds  |
-./spec/raygun/breadcrumbs/store_spec.rb[1:5:8]                   | passed | 0.00053 seconds |
-./spec/raygun/raygun_spec.rb[1:1:1:1:1]                          | passed | 0.10371 seconds |
-./spec/raygun/raygun_spec.rb[1:2:1]                              | passed | 0.00066 seconds |
-./spec/services/apply_whitelist_filter_to_payload_spec.rb[1:1:1] | passed | 0.00027 seconds |
-./spec/services/apply_whitelist_filter_to_payload_spec.rb[1:1:2] | passed | 0.0003 seconds  |
-./spec/services/apply_whitelist_filter_to_payload_spec.rb[1:2:1] | passed | 0.00026 seconds |
-./spec/services/apply_whitelist_filter_to_payload_spec.rb[1:2:2] | passed | 0.00032 seconds |
-./spec/services/apply_whitelist_filter_to_payload_spec.rb[1:2:3] | passed | 0.00028 seconds |
-./spec/services/apply_whitelist_filter_to_payload_spec.rb[1:2:4] | passed | 0.00036 seconds |
-./spec/services/apply_whitelist_filter_to_payload_spec.rb[1:3:1] | passed | 0.00035 seconds |
-./spec/services/apply_whitelist_filter_to_payload_spec.rb[1:4]   | passed | 0.00054 seconds |
