+++ spec/spec_helper.rb
@@ -1,7 +1,6 @@
 require 'rubygems'
-require 'bundler/setup'
 
-require "timecop"
+require 'timecop'
 require 'webmock/rspec'
 
 # Coverage
@@ -10,38 +9,17 @@ require 'webmock/rspec'
   #add_filter '/spec/'
 #end
 
-# This Gem
-require 'raygun'
-
-#Dir[Rails.root.join('spec', 'support', '**', '*.rb')].each { |f| require f }
+require 'support/fake_logger'
 
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
-  config.example_status_persistence_file_path = ".rspec_status"
+  # config.example_status_persistence_file_path = ".rspec_status"
 
   # Disable RSpec exposing methods globally on `Module` and `main`
   # config.disable_monkey_patching!
 
   config.expect_with :rspec do |c|
     c.syntax = :expect
-  end
-end
-
-class FakeLogger
-  def initialize
-    @logger = StringIO.new
-  end
-
-  def info(message)
-    @logger.write(message)
-  end
-
-  def reset
-    @logger.string = ""
-  end
-
-  def get
-    @logger.string
   end
 end
